Rename `spkr` to `PCSpeaker`
