+++ Kernel/Devices/PCSpeaker/spkr.hpp
